+++ docs/md/releasenotes.md
@@ -4,6 +4,7 @@
 
 - `ParticularReferenceValue` sanity checks now take inheritance into consideration, improving call analysis (`T15197`)
 - Prevent missing semicolon leading to an infinite loop in `ClassUtil#internalMethodParameterCount`.
+- Make category 2 CPA taint sources affect only the most significant byte abstract state.
 
 ## Version 9.0 (April 2022)